refactor: register pandas extension at top level
